+++ api/contribution/create.php
@@ -1,0 +1,66 @@
+<?php
+// required headers
+header("Access-Control-Allow-Origin: *");
+header("Content-Type: application/json; charset=UTF-8");
+header("Access-Control-Allow-Methods: POST");
+header("Access-Control-Max-Age: 3600");
+header("Access-Control-Allow-Headers: Content-Type, Access-Control-Allow-Headers, Authorization, X-Requested-With");
+  
+// get database connection
+include_once '../config/database.php';
+  
+// instantiate product object
+include_once '../objects/contributions.php';
+  
+$database = new Database();
+$db = $database->getConnection();
+  
+$contribution = new Contribution($db);
+  
+// get posted data
+$data = json_decode(file_get_contents("php://input"));
+  
+// make sure data is not empty
+if(
+    !empty($data->Name) &&
+    !empty($data->Transaction_no) 
+){
+  
+    // set product property values
+    $contribution->Name = $data->Name;
+    $contribution->Transaction_no = $data->Transaction_no;
+
+
+
+  
+    // create the product
+    if($contribution->create()){
+  
+        // set response code - 201 created
+        http_response_code(201);
+  
+        // tell the user
+        echo json_encode(array("message" => "Contribution was created."));
+    }
+  
+    // if unable to create the product, tell the user
+    else{
+  
+        // set response code - 503 service unavailable
+        http_response_code(503);
+  
+        // tell the user
+        echo json_encode(array("message" => "Unable to create contribution. Choose a unique contribution for the event or check that event is entered correctly."));
+    }
+}
+  
+// tell the user data is incomplete
+else{
+  
+    // set response code - 400 bad request
+    http_response_code(400);
+  
+    // tell the user
+    echo json_encode(array("message" => "Unable to create contribution. Data is incomplete."));
+}
+?>